Removed redundancies and refactored MathUtil
